+++ src/Configuration.scala
@@ -1,7 +1,7 @@
 /*
  This file is part of Configrity.
  
- Foobar is free software: you can redistribute it and/or modify
+ Configrity is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
  the Free Software Foundation, either version 3 of the License, or
  (at your option) any later version.
@@ -105,7 +105,7 @@ object Configuration {
    *  eventually a specified format. By default, the FlatFormat
    *  will be used.
    */
-  def from( s: String, fmt: ImportFormat = defaultFormat ) = 
+  def parse( s: String, fmt: ImportFormat = defaultFormat ) = 
     fmt.fromText( s )
 
   /**