+++ CHANGELOG.md
@@ -1,15 +1,22 @@
 # Changelog #
 
+## Version 0.3.0 - 2011-07-02 ##
+
+### New features ###
+    
+  - Interconversions with java.util.Properties
+  - java.util.Properties Format
+
 ## Version 0.2.0 - 2011-07-01 ##
 
-  ### New features ###
+### New features ###
     
   - FlatFormat
   - From and to string conversion
   . Load and save to files and scala.io.Source
   - Pretty text formatting
 
-  ### Code quality ###
+### Code quality ###
 
   - 100% test coverage with SCCT